+++ cmd2/cmd2.py
@@ -1675,10 +1675,13 @@ class Cmd(cmd.Cmd):
         statement = self.statement_parser.parse_command_only(line)
         return statement.command, statement.args, statement.command_and_args
 
-    def onecmd_plus_hooks(self, line: str) -> bool:
+    def onecmd_plus_hooks(self, line: str, pyscript_bridge_call: bool = False) -> bool:
         """Top-level function called by cmdloop() to handle parsing a line and running the command and all of its hooks.
 
         :param line: line of text read from input
+        :param pyscript_bridge_call: This should only ever be set to True by PyscriptBridge to signify the beginning
+                                     of an app() call in a pyscript. It is used to enable/disable the storage of the
+                                     command's stdout.
         :return: True if cmdloop() should exit, False otherwise
         """
         import datetime
@@ -1718,6 +1721,10 @@ class Cmd(cmd.Cmd):
             try:
                 # Get sigint protection while we set up redirection
                 with self.sigint_protection:
+                    if pyscript_bridge_call:
+                        # Start saving command's stdout at this point
+                        self.stdout.pause_storage = False
+
                     redir_error, saved_state = self._redirect_output(statement)
                     self.cur_pipe_proc_reader = saved_state.pipe_proc_reader
 
@@ -1762,6 +1769,10 @@ class Cmd(cmd.Cmd):
 
                     if not already_redirecting:
                         self.redirecting = False
+
+                    if pyscript_bridge_call:
+                        # Stop saving command's stdout before command finalization hooks run
+                        self.stdout.pause_storage = True
 
         except EmptyStatement:
             # don't do anything, but do allow command finalization hooks to run
@@ -3022,11 +3033,10 @@ class Cmd(cmd.Cmd):
     @with_argparser(py_parser, preserve_quotes=True)
     def do_py(self, args: argparse.Namespace) -> bool:
         """Invoke Python command or shell"""
-        from .pyscript_bridge import PyscriptBridge, CommandResult
+        from .pyscript_bridge import PyscriptBridge
         if self._in_py:
             err = "Recursively entering interactive Python consoles is not allowed."
             self.perror(err, traceback_war=False)
-            self._last_result = CommandResult('', err)
             return False
 
         try: